Add Concatenate Inner Features
