+++ testing/test_zip_data_container.py
@@ -24,7 +24,7 @@ def test_zip_data_container_should_merge_two_data_sources_together():
 def test_zip_data_container_should_merge_multiple_data_sources_together():
     data_inputs_3d, expected_outputs_3d = _create_data_source((10, 10, 2))
     data_inputs_2d, expected_outputs_2d = _create_data_source((10, 10))
-    data_inputs_1d, expected_outputs_1d = _create_data_source((10, ))
+    data_inputs_1d, expected_outputs_1d = _create_data_source((10,))
     data_container_1d = DataContainer(data_inputs=data_inputs_1d, expected_outputs=expected_outputs_1d)
     data_container_2d = DataContainer(data_inputs=data_inputs_2d, expected_outputs=expected_outputs_2d)
     data_container = DataContainer(data_inputs=data_inputs_3d, expected_outputs=expected_outputs_3d)
@@ -39,6 +39,19 @@ def test_zip_data_container_should_merge_multiple_data_sources_together():
     for i, eo in enumerate(zip_data_container.expected_outputs):
         assert np.array_equal(eo[0], expected_outputs_3d[i])
         assert np.array_equal(eo[1], expected_outputs_2d[i])
+
+
+def test_zip_data_container_should_concatenate_inner_features():
+    data_inputs_3d, expected_outputs_3d = _create_data_source((10, 10, 2))
+    data_inputs_2d, expected_outputs_2d = _create_data_source((10, 10))
+    data_container_2d = DataContainer(data_inputs=data_inputs_2d, expected_outputs=expected_outputs_2d)
+    data_container = DataContainer(data_inputs=data_inputs_3d, expected_outputs=expected_outputs_3d)
+
+    zip_data_container = ZipDataContainer.create_from(data_container, data_container_2d)
+    zip_data_container.concatenate_inner_features()
+
+    assert np.array_equal(np.array(zip_data_container.data_inputs)[..., -1], data_container_2d.data_inputs)
+    assert np.array_equal(np.array(zip_data_container.expected_outputs)[..., -1], data_container.data_inputs)
 
 
 def _create_data_source(shape):
